pyproject.toml
[tool.poetry]
name = "nonebot-plugin-ocgbot-v2"
version = "0.2.6"
description = "nonebot-plugin-ocgbot-v2"
authors = ["fireinsect <847954981@qq.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
nonebot2 = ">=2.0.0"
nonebot-adapter-onebot = ">=2.2.3"
requests_html = ">=0.10.0"
Pillow = ">=9.5.0"
httpx = ">=0.20.0,<1.0.0"
numpy = ">=1.23.0"
matplotlib = ">=3.7.1"
nonebot-plugin-localstore = "^0.4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = ["nonebot_plugin_ocgbot_v2"]
builtin_plugins = ["echo"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
